moved IsRedemptionRateWithinSafetyBounds
