+++ x/stakeibc/keeper/invariants.go
@@ -3,9 +3,13 @@ package keeper
 // DONTCOVER
 
 import (
+	"fmt"
+
+	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	epochtypes "github.com/Stride-Labs/stride/v18/x/epochs/types"
+	"github.com/Stride-Labs/stride/v18/x/stakeibc/types"
 )
 
 // RegisterInvariants registers all governance invariants.
@@ -35,4 +39,31 @@ func (k Keeper) AssertStrideAndDayEpochRelationship(ctx sdk.Context) {
 	if dayEpoch.Duration/strideEpoch.Duration != StrideEpochsPerDayEpoch {
 		panic("The stride epoch must be 1/4th the length of the day epoch")
 	}
+}
+
+// safety check: ensure the redemption rate is NOT below our min safety threshold && NOT above our max safety threshold on host zone
+func (k Keeper) IsRedemptionRateWithinSafetyBounds(ctx sdk.Context, zone types.HostZone) (bool, error) {
+	// Get the wide bounds
+	minSafetyThreshold, maxSafetyThreshold := k.GetOuterSafetyBounds(ctx, zone)
+
+	redemptionRate := zone.RedemptionRate
+
+	if redemptionRate.LT(minSafetyThreshold) || redemptionRate.GT(maxSafetyThreshold) {
+		errMsg := fmt.Sprintf("IsRedemptionRateWithinSafetyBounds check failed %s is outside safety bounds [%s, %s]", redemptionRate.String(), minSafetyThreshold.String(), maxSafetyThreshold.String())
+		k.Logger(ctx).Error(errMsg)
+		return false, errorsmod.Wrapf(types.ErrRedemptionRateOutsideSafetyBounds, errMsg)
+	}
+
+	// Verify the redemption rate is within the inner safety bounds
+	// The inner safety bounds should always be within the safety bounds, but
+	// the redundancy above is cheap.
+	// There is also one scenario where the outer bounds go within the inner bounds - if they're updated as part of a param change proposal.
+	minInnerSafetyThreshold, maxInnerSafetyThreshold := k.GetInnerSafetyBounds(ctx, zone)
+	if redemptionRate.LT(minInnerSafetyThreshold) || redemptionRate.GT(maxInnerSafetyThreshold) {
+		errMsg := fmt.Sprintf("IsRedemptionRateWithinSafetyBounds check failed %s is outside inner safety bounds [%s, %s]", redemptionRate.String(), minInnerSafetyThreshold.String(), maxInnerSafetyThreshold.String())
+		k.Logger(ctx).Error(errMsg)
+		return false, errorsmod.Wrapf(types.ErrRedemptionRateOutsideSafetyBounds, errMsg)
+	}
+
+	return true, nil
 }
